Move files to providers module
